+++ pdpc_decisions/save_file.py
@@ -4,8 +4,7 @@ import csv
 import logging
 from typing import List
 
-from pdpc_decisions.pdpcdecision import Options
-from pdpc_decisions.scraper import PDPCDecisionItem
+from pdpc_decisions.classes import Options, PDPCDecisionItem
 
 logger = logging.getLogger(__name__)
